+++ src/components/CostumeCard/CostumeCard.jsx
@@ -3,7 +3,10 @@ import { string, bool, func, shape } from 'prop-types';
 import { FaLink } from 'react-icons/fa6';
 import { TbPhotoExclamation } from 'react-icons/tb'; // 대체 아이콘
 import S from './CostumeCard.module.scss';
+import LazyLoad from 'react-lazyload';
 import { Loader } from '@/components';
+import { useLocation } from 'react-router-dom';
+import clsx from 'clsx';
 
 const CostumeCard = ({ record, imageUrl, isLiked, onLikeToggle }) => {
   const { costumeTitle, costumeBrand, costumeLink } = record;
@@ -24,7 +27,7 @@ const CostumeCard = ({ record, imageUrl, isLiked, onLikeToggle }) => {
   }, [imageUrl, isLoading]);
 
   const handleImageLoad = () => {
-    console.log(`Image loaded successfully: ${validImageUrl}`);
+    // console.log(`Image loaded successfully: ${validImageUrl}`);
     setIsLoading(false); // 로딩 종료
     setFileStatus('loaded');
   };
@@ -47,6 +50,12 @@ const CostumeCard = ({ record, imageUrl, isLiked, onLikeToggle }) => {
     }
   };
 
+  // 현재 경로가 좋아요 페이지면 .image 에 .imageInLikePage 클래스 추가
+  const location = useLocation();
+  const imageClass = clsx(S.image, {
+    [S.imageInLikePage]: location.pathname === '/liked',
+  });
+
   return (
     <div className={S.card}>
       {/* 좋아요 버튼 */}
@@ -57,21 +66,26 @@ const CostumeCard = ({ record, imageUrl, isLiked, onLikeToggle }) => {
           className={S.likeIcon}
         />
       </button>
-
       {/* 상단의 제품 이미지 또는 에러 아이콘 */}
       <div className={S.imageWrapper}>
         {renderStatus()}
-        {validImageUrl && !imageError ? (
-          <img
-            src={validImageUrl}
-            alt={costumeTitle}
-            className={S.image}
-            onLoad={handleImageLoad}
-            onError={handleImageError}
-          />
-        ) : (
-          <TbPhotoExclamation className={S.errorIcon} /> // 이미지 로딩 실패 또는 경로가 없는 경우 아이콘 표시
-        )}
+        <LazyLoad
+          height={175} // 이미지 높이를 175px로 설정
+          offset={50} // 스크롤 100px 전에 로딩 시작
+          placeholder={<Loader />} // 로딩 중 보여줄 컴포넌트
+        >
+          {validImageUrl && !imageError ? (
+            <img
+              src={validImageUrl}
+              alt={costumeTitle}
+              className={imageClass}
+              onLoad={handleImageLoad}
+              onError={handleImageError}
+            />
+          ) : (
+            <TbPhotoExclamation className={S.errorIcon} /> // 이미지 로딩 실패 또는 경로가 없는 경우 아이콘 표시
+          )}
+        </LazyLoad>
       </div>
 
       {/* 제품 정보 */}
