+++ src/components/CostumeCard/CostumeCard.module.scss
@@ -1,18 +1,14 @@
 @import '@/styles/utils/_utils.scss';
 .card {
+  flex: 1 1 40%;
   display: flex;
   flex-direction: column;
-
-  width: 150px;
-  min-height: 315px;
-
-  border: 1px solid #ddd;
-  border-radius: 10px;
+  position: relative; /* 하트 버튼이 카드 내에서 위치하도록 설정 */
+  max-width: 138px;
+  // min-height: 292px;
   overflow: hidden;
-  box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.1);
   background-color: #fff;
   transition: transform 0.2s ease-in-out;
-  position: relative; /* 하트 버튼이 카드 내에서 위치하도록 설정 */
 
   &:hover {
     transform: translateY(-5px);
@@ -20,8 +16,8 @@
 }
 
 .imageWrapper {
-  width: 100%;
-  height: 100%;
+  // width: 100%;
+  // height: 100%;
 
   position: relative;
   display: flex;
@@ -29,11 +25,13 @@
   align-items: center;
 
   .image {
-    margin-top: 30px;
-
     width: 100%;
     height: 175px;
     object-fit: cover;
+  }
+
+  .imageInLikePage{
+    height:120px; /* 좋아요 페이지에 적용되는 이미지 높이 */
   }
 
   .errorIcon {
@@ -44,25 +42,33 @@
 }
 
 .infoWrapper {
-  padding: 1rem;
+  font-size:14px;
+  padding: 0.6rem 0.6rem 0 0;
 
   .title {
+    margin-bottom:6px;
     font-weight: 700;
   }
 
   .description {
-    @include ellipsis(1, 1.2em); /* 한 줄만 표시하고 나머지는 숨김 */
+    @include ellipsis(2, 1.2em); /* 두 줄만 표시하고 나머지는 숨김 */
     text-overflow: ellipsis;
     overflow: hidden;
-    white-space: nowrap; /* 텍스트를 한 줄로 제한 */
+    margin-bottom:6px;
   }
 
   .link {
+    display:flex;
+    align-items: baseline;
     text-decoration: none;
     color: var(--black);
     font-weight: 600;
-    &:hover {
-      cursor: pointer;
+    cursor: pointer;
+
+    img, svg{
+      display:inline-block;
+      margin-right:2px;
+      padding-top: 3px;
     }
   }
 }
@@ -71,7 +77,8 @@
   position: absolute; /* imageWrapper 기준으로 하트 위치 */
   top: 10px;
   right: 10px;
-  width: 10%;
+  width: 14%;
+  min-width:20px;
   background: none;
   border: none;
   cursor: pointer;
