--- src/components/CostumeCard_hjm/CostumeCard.module.scss
@@ -1,91 +1,0 @@
-@import '@/styles/utils/_utils.scss';
-.card {
-  flex: 1 1 40%;
-  display: flex;
-  flex-direction: column;
-  position: relative; /* 하트 버튼이 카드 내에서 위치하도록 설정 */
-  max-width: 138px;
-  // min-height: 292px;
-  overflow: hidden;
-  background-color: #fff;
-  transition: transform 0.2s ease-in-out;
-
-  &:hover {
-    transform: translateY(-5px);
-  }
-}
-
-.imageWrapper {
-  // width: 100%;
-  // height: 100%;
-
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  .image {
-    width: 100%;
-    height: 175px;
-    object-fit: cover;
-  }
-
-  .imageInLikePage{
-    height:120px; /* 좋아요 페이지에 적용되는 이미지 높이 */
-  }
-
-  .errorIcon {
-    font-size: 3rem; /* 에러 아이콘의 크기 */
-    color: var(--primary-color); /* 에러 아이콘의 색상 */
-    object-fit: contain;
-  }
-}
-
-.infoWrapper {
-  font-size:14px;
-  padding: 0.6rem 0.6rem 0 0;
-
-  .title {
-    margin-bottom:6px;
-    font-weight: 700;
-  }
-
-  .description {
-    @include ellipsis(2, 1.2em); /* 두 줄만 표시하고 나머지는 숨김 */
-    text-overflow: ellipsis;
-    overflow: hidden;
-    margin-bottom:6px;
-  }
-
-  .link {
-    display:flex;
-    align-items: baseline;
-    text-decoration: none;
-    color: var(--black);
-    font-weight: 600;
-    cursor: pointer;
-
-    img, svg{
-      display:inline-block;
-      margin-right:2px;
-      padding-top: 3px;
-    }
-  }
-}
-
-.likeButton {
-  position: absolute; /* imageWrapper 기준으로 하트 위치 */
-  top: 10px;
-  right: 10px;
-  width: 14%;
-  min-width:20px;
-  background: none;
-  border: none;
-  cursor: pointer;
-  z-index: 10;
-  transition: transform 0.2s ease-in-out;
-
-  &:hover {
-    transform: scale(1.1);
-  }
-}
